% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonode4R.R
\docType{package}
\name{geonode4R}
\alias{geonode4R}
\alias{geonode4R-package}
\title{GeoNode REST API R Interface}
\description{
Provides an R interface to the GeoNode API, allowing to 
upload and publish metadata and data in GeoNode. For more information 
about the GeoNode API, see <https://docs.geonode.org/en/master/devel/api/V2/index.html>
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
